refactor: rename Request interface to Credentials and remove unused Results class
- Rename 'Request' interface to 'Credentials' in Base\Domain\Contract\Authentication

- Rename Request.php to Credentials.php, update all typehints and imports

- Delete unused Results.php class from the same namespace

- Update test files to reference Credentials instead of Request
