+++ test/User/Authentication/Domain/Strategy/ChallengeCodeStrategyTest.php
@@ -16,14 +16,14 @@ namespace Webify\Test\User\Authentication\Domain\Strategy;
 use PHPUnit\Framework\Attributes\{AllowMockObjectsWithoutExpectations, CoversClass, CoversMethod, Test};
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Webify\Base\Domain\Contract\Authentication\{Request, UserCredentials};
+use Webify\Base\Domain\Contract\Authentication\{Credentials, UserCredentials};
 use Webify\Base\Domain\Event\DomainEventPublisherInterface;
 use Webify\Base\Domain\Service\UlidGeneratorInterface;
 use Webify\User\Authentication\Domain\Entity\Challenge;
 use Webify\User\Authentication\Domain\Repository\ChallengeRepositoryInterface;
 use Webify\User\Authentication\Domain\Service\{ChallengeSecretDeliveryInterface, IssueChallenge};
 use Webify\User\Authentication\Domain\Strategy\ChallengeCodeStrategy;
-use Webify\User\Authentication\Domain\ValueObject\{UserId};
+use Webify\User\Authentication\Domain\ValueObject\UserId;
 
 /**
  * Tests for the ChallengeCodeStrategy.
@@ -102,7 +102,7 @@ final class ChallengeCodeStrategyTest extends TestCase
 	#[Test]
 	public function testIsSupportedWithMatchingIdentifier(): void
 	{
-		$request = $this->createStub(Request::class);
+		$request = $this->createStub(Credentials::class);
 		$request
 			->method('getStrategyIdentifier')
 			->willReturn('challenge_code')
@@ -117,7 +117,7 @@ final class ChallengeCodeStrategyTest extends TestCase
 	#[Test]
 	public function testIsSupportedWithNonMatchingIdentifier(): void
 	{
-		$request = $this->createStub(Request::class);
+		$request = $this->createStub(Credentials::class);
 		$request
 			->method('getStrategyIdentifier')
 			->willReturn('challenge_token')
@@ -141,7 +141,7 @@ final class ChallengeCodeStrategyTest extends TestCase
 			status: 'active',
 			isTwoFactorEnabled: true
 		);
-		$request = $this->createStub(Request::class);
+		$request = $this->createStub(Credentials::class);
 
 		$this->challengeRepository
 			->expects($this->once())
